app/requirements.txt
attrs==23.1.0
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
outcome==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==1.26.16
wsproto==1.2.0
